+++ src/components/ui/radio-group/RadioGroup.vue
@@ -1,8 +1,8 @@
 <script setup lang="ts">
-import type { RadioGroupRootEmits, RadioGroupRootProps } from 'reka-ui'
 import { RadioGroupRoot, useForwardPropsEmits } from 'reka-ui'
-import type { HTMLAttributes } from 'vue'
 import { cn } from '@/lib/utils'
+import type { RadioGroupRootEmits, RadioGroupRootProps } from 'reka-ui'
+import type { HTMLAttributes } from 'vue'
 
 const props = defineProps<RadioGroupRootProps & { class?: HTMLAttributes['class'] }>()
 const emits = defineEmits<RadioGroupRootEmits>()